requirements.txt
Django==3.2.12
django-crispy-forms==1.12.0
django-impersonate==1.7.3
django-chunked-upload==2.0.0
django-oauth-toolkit==1.5.0
django-auth-ldap==3.0.0
celery==5.2.3
redis==4.1.2
requests==2.27.1
kubernetes==20.13.0
ruamel.yaml==0.17.21
passlib==1.7.4
Jinja2==3.1.1
packaging==21.3
jsonpatch==1.32
gunicorn==20.1.0
whitenoise==5.3.0
